requirements.txt
discord.py
python-dotenv
pandas
openpyxl
asyncpg
xlsxwriter
wavelink
yt-dlp
ffmpeg
PyNaCl
matplotlib
aiohttp